IOS-7560: Rename files, improve project file structure
