+++ Source/Charts/Charts/ColorSplitLineChart/ColorSplitLineChartSegmentAppearance.swift
@@ -1,5 +1,5 @@
 //
-//  LineChartContainerViewControllerSegmentAppearance.swift
+//  ColorSplitLineChartSegmentAppearance.swift
 //  Tangem
 //
 //  Created by Andrey Fedorov on 22.08.2024.
@@ -9,7 +9,7 @@
 import Foundation
 import QuartzCore
 
-public struct LineChartContainerViewControllerSegmentAppearance {
+public struct ColorSplitLineChartSegmentAppearance {
     public struct Gradient {
         public init(
             colors: [NSUIColor],
